backend/photos/requirements.txt
boto3==1.34.51
psycopg2-binary==2.9.9
